complete change on roles: delete <isadmin> for <roles class>

+++ interface/body/header.php
@@ -23,36 +23,71 @@ $getPage = Lib::Sanitize($_GET['page']);
         Interface
     </div>
     <!-- Nav Item - Pages Collapse Menu -->
-    <?php if ($current_user->getGrade() == true) { ?>
-    <li class="nav-item  <?php if ($getPage == 'add_user' || $getPage == 'edit_user' || $getPage == 'delete_user'
-    || $getPage == 'generate_auth_token' || $getPage == 'delete_auth_token' || $getPage == 'logs') echo 'active'; ?>">
+    <?php if ($current_user->asAuthorizationToAccess(array(
+        "Manage Users",
+        "Manage Tokens",
+        "Manage Roles",
+        "Consult Logs",
+        "Project Administration",
+    ))) { ?>
+    <li class="nav-item">
         <a class="nav-link collapsed" href="#" data-toggle="collapse" data-target="#collapseTwo" aria-expanded="true" aria-controls="collapseTwo">
             <i class="fas fa-fw fa-cog"></i>
             <span>Users Config</span>
         </a>
         <div id="collapseTwo" class="collapse" aria-labelledby="headingTwo" data-parent="#accordionSidebar">
             <div class="bg-white py-2 collapse-inner rounded">
+                <?php
+                if ($current_user->asAuthorizationToAccess("Manage Users")) {
+                ?>
                 <h6 class="collapse-header">Manage Users</h6>
                 <a class="collapse-item" href="index.php?page=add_user">Register new user</a>
                 <a class="collapse-item" href="index.php?page=edit_user_roles">Edit users roles</a>
                 <a class="collapse-item" href="index.php?page=delete_user_roles">Delete users roles</a>
                 <a class="collapse-item" href="index.php?page=delete_user">Delete users</a>
+                <?php
+                }
+
+                if ($current_user->asAuthorizationToAccess("Manage Tokens")) {
+                ?>
                 <h6 class="collapse-header">Manage Tokens</h6>
                 <a class="collapse-item" href="index.php?page=generate_auth_token">Generate auth tokens</a>
                 <a class="collapse-item" href="index.php?page=delete_auth_token">Delete auth tokens</a>
+                <?php
+                }
+
+                if ($current_user->asAuthorizationToAccess("Manage Roles")) {
+                ?>
                 <h6 class="collapse-header">Manage Roles</h6>
                 <a class="collapse-item" href="index.php?page=add_roles">Add roles</a>
                 <a class="collapse-item" href="index.php?page=edit_roles">Edit Roles</a>
                 <a class="collapse-item" href="index.php?page=delete_roles">Delete roles</a>
+                <?php
+                }
+
+                if ($current_user->asAuthorizationToAccess("Consult Logs")) {
+                ?>
                 <h6 class="collapse-header">Other</h6>
                 <a class="collapse-item" href="index.php?page=logs">View Logs</a>
+                <?php
+                }
+
+                if ($current_user->asAuthorizationToAccess("Project Administration")) {
+                ?>
                 <h6 class="collapse-header">Project Administration</h6>
                 <a class="collapse-item" href="">Consult reports</a>
+                <?php
+                }
+                ?>
             </div>
         </div>
     </li>
-<?php } ?>
-    <li class="nav-item <?php if ($getPage == 'roadmap' || $getPage == 'unitest' || $getPage == 'benchmark' || $getPage == 'chat') echo 'active'; ?>">
+    <?php
+    }
+
+    if ($current_user->asAuthorizationToAccess("Utilities")) {
+        ?>
+    <li class="nav-item">
         <a class="nav-link collapsed" href="#" data-toggle="collapse" data-target="#collapseUtilities" aria-expanded="true" aria-controls="collapseUtilities">
             <i class="fas fa-fw fa-wrench"></i>
             <span>Utilities</span>
@@ -67,6 +102,11 @@ $getPage = Lib::Sanitize($_GET['page']);
             </div>
         </div>
     </li>
+    <?php
+    }
+
+    if ($current_user->asAuthorizationToAccess("Calendar")) {
+    ?>
     <li class="nav-item">
         <a class="nav-link collapsed" href="#" data-toggle="collapse" data-target="#collapseCalendar" aria-expanded="true" aria-controls="collapseUtilities">
             <i class="fas fa-fw fa-calendar-alt"></i>
@@ -79,6 +119,14 @@ $getPage = Lib::Sanitize($_GET['page']);
             </div>
         </div>
     </li>
+    <?php
+    }
+
+    if ($current_user->asAuthorizationToAccess(array(
+        "Manage Article",
+        "Manage Games"
+    ))) {
+    ?>
     <!-- Divider -->
     <hr class="sidebar-divider">
     <!-- Heading -->
@@ -86,26 +134,45 @@ $getPage = Lib::Sanitize($_GET['page']);
         Website
     </div>
     <!-- Nav Item - Pages Collapse Menu -->
-    <li class="nav-item <?php if ($getPage == 'add_article' || $getPage == 'edit_article' || $getPage == 'delete_article'
-    || $getPage == 'add_game' || $getPage == 'edit_game' || $getPage == 'delete_game') echo 'active'; ?>">
+    <?php
+    if ($current_user->asAuthorizationToAccess(array(
+        "Manage Article",
+        "Manage Games"
+    ))) {
+    ?>
+    <li class="nav-item">
         <a class="nav-link collapsed" href="#" data-toggle="collapse" data-target="#leaflighted" aria-expanded="true" aria-controls="collapsePages">
             <i class="fas fa-fw fa-folder"></i>
             <span>Leaflighted</span>
         </a>
         <div id="leaflighted" class="collapse" aria-labelledby="headingPages" data-parent="#accordionSidebar">
             <div class="bg-white py-2 collapse-inner rounded">
+                <?php
+                if ($current_user->asAuthorizationToAccess("Manage Article")) {
+                ?>
                 <h6 class="collapse-header">Manage Article:</h6>
                 <a class="collapse-item" href="index.php?page=add_article">Add Article</a>
                 <a class="collapse-item" href="index.php?page=edit_article">Edit Article</a>
                 <a class="collapse-item" href="index.php?page=delete_article">Delete Article</a>
+                <?php
+                }
+
+                if ($current_user->asAuthorizationToAccess("Manage Games")) {
+                ?>
                 <div class="collapse-divider"></div>
                 <h6 class="collapse-header">Manage games:</h6>
                 <a class="collapse-item" href="index.php?page=add_game">Add game</a>
                 <a class="collapse-item" href="index.php?page=edit_game">Edit game</a>
                 <a class="collapse-item" href="index.php?page=delete_game">Delete game</a>
+                <?php
+                }
+                ?>
             </div>
         </div>
     </li>
+        <?php
+        }
+        ?>
     <li class="nav-item">
         <a class="nav-link collapsed" href="#" data-toggle="collapse" data-target="#rite" aria-expanded="true" aria-controls="collapsePages">
             <i class="fas fa-fw fa-folder"></i>
@@ -117,6 +184,9 @@ $getPage = Lib::Sanitize($_GET['page']);
             </div>
         </div>
     </li>
+    <?php
+    }
+    ?>
 <hr class="sidebar-divider d-none d-md-block">
 <div class="text-center d-none d-md-inline">
     <button class="rounded-circle border-0" id="sidebarToggle"></button>
